mm: make vm_munmap killable
Almost all current users of vm_munmap are ignoring the return value and so
they do not handle potential error.  This means that some VMAs might stay
behind.  This patch doesn't try to solve those potential problems.  Quite
contrary it adds a new failure mode by using down_write_killable in
vm_munmap.  This should be safer than other failure modes, though, because
the process is guaranteed to die as soon as it leaves the kernel and
exit_mmap will clean the whole address space.

This will help in the OOM conditions when the oom victim might be stuck
waiting for the mmap_sem for write which in turn can block oom_reaper
which relies on the mmap_sem for read to make a forward progress and
reclaim the address space of the victim.

Signed-off-by: Michal Hocko <mhocko@suse.com>
Cc: Oleg Nesterov <oleg@redhat.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Konstantin Khlebnikov <koct9i@gmail.com>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: Alexander Viro <viro@zeniv.linux.org.uk>
Cc: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>