mm: add config option to select the initial overcommit mode
Currently the initial value of the overcommit mode is OVERCOMMIT_GUESS.
However, on embedded systems it is usually better to disable overcommit to
avoid waking up the OOM-killer and its well known undesirable
side-effects.

This config option allows to setup the initial overcommit mode to any of
the 3 available values, OVERCOMMIT_GUESS (which remains as default),
OVERCOMMIT_ALWAYS and OVERCOMMIT_NEVER.  The overcommit mode can still be
changed thru sysctl after the system boots up.

This config option depends on CONFIG_EXPERT.
This patch does not introduces functional changes.

Link: http://lkml.kernel.org/r/5731CC6E.3080807@laposte.net
Signed-off-by: Sebastian Frias <sf84@laposte.net>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>