m32r: fix build failure
m32r allmodconfig build was failing with the error: ERROR:
"smp_flush_cache_all" [drivers/misc/lkdtm.ko] undefined!

lkdtm driver at drivers/misc is using flush_icache_range() which for m32r
is defined as smp_flush_cache_all() if CONFIG_SMP is defined.  But as
smp_flush_cache_all() was not exported so the build was failing with the
error of undefined symbol.

Link: http://lkml.kernel.org/r/1464001182-8329-1-git-send-email-sudipm.mukherjee@gmail.com
Signed-off-by: Sudip Mukherjee <sudip.mukherjee@codethink.co.uk>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>