dm crypt: fix error with too large bios
When dm-crypt processes writes, it allocates a new bio in
crypt_alloc_buffer().  The bio is allocated from a bio_set and it can
have at most BIO_MAX_PAGES vector entries, however the incoming bio
can be larger if it was allocated by other means.  For example, bcache
creates bios larger than BIO_MAX_PAGES (leaning on late bio-splitting).
But if the incoming bio is larger than BIO_MAX_PAGES bio_alloc_bioset()
fails and an error is returned.

To avoid the error, we test for too large bio in crypt_map() and use
dm_accept_partial_bio() to split the bio.  dm_accept_partial_bio()
trims the current bio to the desired size and requests that DM core
send another bio with the rest of the data.

Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>
Cc: stable@vger.kernel.org	# v3.16+
Signed-off-by: Mike Snitzer <snitzer@redhat.com>