PM / Runtime: Defer resuming of devices in pm_runtime_force_resume()
When the pm_runtime_force_suspend|resume() helpers were invented, we
still had CONFIG_PM_RUNTIME and CONFIG_PM_SLEEP as separate Kconfig
options.

To make sure these helpers worked for all combinations and without
introducing too much of complexity, the device was always resumed
in pm_runtime_force_resume().

More precisely, when CONFIG_PM_SLEEP was set and CONFIG_PM_RUNTIME
was unset, we needed to resume the device as the subsystem/driver
couldn't rely on using runtime PM to do it.

As the CONFIG_PM_RUNTIME option was merged into CONFIG_PM a while
ago, it removed this combination, of using CONFIG_PM_SLEEP without
the earlier CONFIG_PM_RUNTIME.

For this reason we can now rely on the subsystem/driver to use
runtime PM to resume the device, instead of forcing that to be done
in all cases. In other words, defer the runtime resume to a later
point when it's actually needed.

Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Reviewed-by: Kevin Hilman <khilman@baylibre.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>