ARM/PCI: Remove arch-specific pcibios_enable_device()
pci_enable_resources() fails unless device BARs are assigned and inserted
in the resource tree.  If the BAR resources are fine, it turns on the
PCI_COMMAND_IO and PCI_COMMAND_MEMORY bits if necessary.

On systems with PCI_PROBE_ONLY set, we rely on BAR assignments from
firmware, but previously we did not insert them into the resource tree, so
we skipped pci_enable_resources() because it would always fail.

Now that we *do* insert resources even when PCI_PROBE_ONLY is set, we no
longer need the arm-specific pcibios_enable_device().  Remove it so we
use the generic version.

[bhelgaas: changelog]
Signed-off-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
Acked-by: Will Deacon <will.deacon@arm.com>
CC: Arnd Bergmann <arnd@arndb.de>
CC: Russell King <linux@arm.linux.org.uk>