kconfig.h: use __is_defined() to check if MODULE is defined
The macro MODULE is not a config option, it is a per-file build option.
So, config_enabled(MODULE) is not sensible.  (There is another case in
include/linux/export.h, where config_enabled() is used against a
non-config option.)

This commit renames some macros in include/linux/kconfig.h for the use for
non-config macros and replaces config_enabled(MODULE) with
__is_defined(MODULE).

I am keeping config_enabled() because it is still referenced from some
places, but I expect it would be deprecated in the future.

Link: http://lkml.kernel.org/r/1465883938-25194-1-git-send-email-yamada.masahiro@socionext.com
Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Cc: Rusty Russell <rusty@rustcorp.com.au>
Cc: Michal Simek <michal.simek@xilinx.com>
Cc: Arnd Bergmann <arnd@arndb.de>
Cc: Michal Marek <mmarek@suse.cz>
Cc: Nicolas Pitre <nicolas.pitre@linaro.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>