HACK: ARM: smccc-call: Use R12 to route secure monitor calls
We signal the secure monitor that our SMC is intended for the TEE or not
with the value in r12. We also need to save additional registers in the
case a compatible TEE is not loaded, as this may clobber these registers.

Signed-off-by: Harinarayan Bhatta <harinarayan@ti.com>
Signed-off-by: Andrew F. Davis <afd@ti.com>
(cherry picked from commit 00f19d3e5347c2992fe0de52cabe5d379673973d)
Signed-off-by: Dan Murphy <dmurphy@ti.com>