mmc: host: omap_hsmmc: keep CON_PADEN always enabled
There is a delay in enumeration of MMC on some DRA7 and AM57 platforms.
Rootcause of this behavior is that there is a delay between setting PADEN
and the time when the status is reflected on the IO lines.
A simple fix is to add a delay in the code after setting the PADEN bit.

The exact delay required is being worked and will be updated in TRM also a
accurate SW fix once the investigation completes.

If PADEN is always set then there’s no delay required when checking for busy signal.
So, set the PADEN in probe() to ensure it’s always enabled.

This fix helps in mounting android system and vendor partition on
DRA76X-EVM and AM572X-GPEVM and per early_mount (first stage init mount)
Ref: https://source.android.com/devices/architecture/kernel/modular-kernels#mounting-partitions-early-first-stage-mount

Signed-off-by: Ravikumar Kattekola <rk@ti.com>
Signed-off-by: Vishal Mahaveer <vishalm@ti.com>
[praneeth@ti.com: forward port from 4.4 to 4.9 and update commitmsg]
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>