ti-config-fragments: Add a config fragment for presil
Adding a config fragment for presil pruning.  The kernel
for presil needs to be compact to reduce load and boot time.

The flags contained in this fragment were added based on the
presil_defconfig from commit the presil_defconfig in the arm64
configs directory in commit

arch: arm64: configs: Add a presilicon regular and a mini defconfig

Signed-off-by: Dan Murphy <dmurphy@ti.com>