drm/omap: dss7: Add OLDI implementation
Add OLDI implementation for DSS7 as oldi7.c driver. The actual OLDI HW
implementation is in dss7.c because the feature is part of DSS7
IP. However, the configuration is handled by oldi7.c. oldi7 driver is
almost identical to dpi7 drive. There one new required property, which
is also used to determine if a port is an OLDI or DPI port.

Signed-off-by: Jyri Sarha <jsarha@ti.com>