drm/omap: Add TI DSS7 driver
Add support for DSS7 IP on TI AM6x SoC.

DSS7 is an evolution of the OMAP DSS (DSS2) and DSS6. OMAP DSS has
been quite similar from OMAP2 to OMAP5 (including DRA7 and AM5 which
have basically the same IP as OMAP5). In DSS6 lots of things have been
restructured and DSS7 is a next step of evolution from K2G DSS6,
basically adding more more features, planes and video ports.

DSS7 and DSS6 can still be considered as being the same family as
DSS2. The features offered are mostly the same, and the model how
things happen and work are the same. What's changed are all the
registers and how the functionalities in the IP are divided. This
makes unifying their backend drivers virtually impossible.

DSS6 and DSS7 are very close relatives and for instance plane and
video port registers are almost exact match. On the other hand the
common registers have been been rearranged, the way the interrupts
work is quite different, and there is no separation between DSS and
DISPC IPs anymore. All this makes it unpractical to unify the drivers, but
in the long run some code sharing between them may make sense.

This patch adds a new DSS driver for DSS7. The DSS2's panel and encoder
drivers and also omapdrm driver can be used with DSS7.

Signed-off-by: Jyri Sarha <jsarha@ti.com>