drm/omap: Hide "cast from pointer to integer of different size"
Get rid of "warning: cast from pointer to integer of different
size". The pointer in 64-bit system is 64-bit. Casting first to
integer of matching integer size (ptrdiff_t) and only then to final
uint32_t makes the compiler happy. The "pointer" is really not a
pointer at all so there is no risk of overflow.

Signed-off-by: Jyri Sarha <jsarha@ti.com>