drivers: gpu: Add PowerVR SGX544 driver
This is the kernel side driver for the PowerVR SGX544 on the AM57x. Its
basic job is to handle passing buffers to the GPU and kicking it as
needed. All the "magic" happens in the userspace driver and the firmware
running in the GPU, this driver simply facilitates their communication.

This driver is currently an almost direct copy/paste from the DDK 1.17
and so is specific to rev 116, full of whitespace problems, and many
other issues like that (don't run checkpatch on this unless you have
a fire extinguisher next to your computer).

I will pick away at these issues as I can on the DDK side and move
the changes over to here. In this way the DDK repo should be considered
as the upstream for this code.

Signed-off-by: Andrew F. Davis <afd@ti.com>