arch: arm64: configs: Remove presil and presil_mini defconfigs
The Kconfig fragments using the defconfig builder script (both
present in ti_config_fragments folder) is the preferred way of
building the Linux kernel for DRA82x pre-silicon platforms. Delete
the previously added presil and presil_mini defconfig files so as
to not confuse developers/integrators to choose the right method
and also to avoid synchronization issues between the two config
files.

Signed-off-by: Suman Anna <s-anna@ti.com>